+++ packages/test-factory/src/test-utils.ts
@@ -22,7 +22,7 @@ export const BaseTable = createBaseTable({
   columnTypes: (t) => ({
     ...t,
     text: (min = 0, max = Infinity) => t.text(min, max),
-    timestamp: () => t.timestamp().asNumber(),
+    timestampWithoutTimeZone: () => t.timestampWithoutTimeZone().asNumber(),
   }),
 });