+++ packages/rake-db/CHANGELOG.md
@@ -1,5 +1,14 @@
 # rake-db
 
+## 2.4.40
+
+### Patch Changes
+
+- Favor timestamptz over timestamp
+- Updated dependencies
+  - orchid-core@0.1.16
+  - pqb@0.10.33
+
 ## 2.4.39
 
 ### Patch Changes