+++ packages/test-factory/CHANGELOG.md
@@ -1,5 +1,16 @@
 # orchid-orm-test-factory
 
+## 0.3.38
+
+### Patch Changes
+
+- Favor timestamptz over timestamp
+- Updated dependencies
+  - orchid-core@0.1.16
+  - pqb@0.10.33
+  - orchid-orm-schema-to-zod@0.3.34
+  - orchid-orm@1.6.37
+
 ## 0.3.37
 
 ### Patch Changes